Rename HomepageBuilderEditModePreview to EditModePreview
